requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.1
django-ckeditor==6.7.0
django-ckeditor-5==0.2.10
django-filter==23.3
django-js-asset==2.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
environs==9.5.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
openapi-codec==1.3.2
packaging==23.2
Pillow==10.1.0
PyJWT==2.8.0
pypng==0.20220715.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
setuptools==68.2.2
simplejson==3.19.2
sqlparse==0.4.4
stripe==7.3.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
